+++ backend/indexing/iteminfo/conditions.go
@@ -20,7 +20,7 @@ var AllFiletypeOptions = []string{
 	"text",
 }
 
-var PdfConvertable = []string{
+var MuPdfConvertable = []string{
 	".pdf",  // PDF
 	".xps",  // XPS
 	".epub", // EPUB
@@ -37,6 +37,7 @@ var PdfConvertable = []string{
 	".xlsx", // XLSX
 	".hwp",  // HWP
 	".hwp",  // HWPX
+	".md",   // Markdown
 }
 
 // Known bundle-style extensions that are technically directories but treated as files
@@ -375,12 +376,15 @@ func hasBundleExtension(name string) bool {
 	return false
 }
 
-func HasDocConvertableExtension(name string) bool {
-	if !settings.Config.Server.PdfAvailable {
+func HasDocConvertableExtension(name, mimetype string) bool {
+	if !settings.Config.Server.MuPdfAvailable {
 		return false
 	}
+	if strings.HasPrefix(mimetype, "text") {
+		return true
+	}
 	ext := strings.ToLower(filepath.Ext(name))
-	for _, e := range PdfConvertable {
+	for _, e := range MuPdfConvertable {
 		if ext == e {
 			return true
 		}